requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cos-python-sdk-v5==1.8.0
dicttoxml==1.7.4
Django==1.11.29
django-redis==4.11.0
idna==2.9
Naked==0.1.31
Pillow==7.1.2
pycryptodome==3.9.7
PyMySQL==0.9.3
pytz==2020.1
PyYAML==5.3.1
qcloud-python-sts==3.0.3
qcloudsms-py==0.1.4
redis==3.5.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
urllib3==1.25.9
